Allow custom Meson CLI path
